requirements.txt
azure-cosmos>=4.6.0
google-cloud-spanner>=3.46.0
google-cloud-firestore>=2.16.0
google-cloud-storage>=2.16.0
PyYAML>=6.0.1
pymongo>=4.7.0
cassandra-driver>=3.29.0
pytest>=8.0.0
